refactor: request 를 controller 레이어로 올리면서, 공통 의존성 제거
