+++ build.gradle
@@ -22,7 +22,6 @@ allprojects {
 
     dependencies {
         implementation 'org.springframework:spring-context'
-        implementation 'org.springframework.boot:spring-boot-starter-validation'
         implementation 'org.springframework.boot:spring-boot-starter-web'
