+++ domain/src/main/java/com/sparcs/teamf/question/dto/PageResponse.java
@@ -1,4 +1,5 @@
 package com.sparcs.teamf.question.dto;
+//todo: 모든 response 도 controller 레이어로 올린다
 
 public record PageResponse(long pageId) {
