+++ web/src/main/java/com/sparcs/teamf/api/auth/dto/SignupRequest.java
@@ -1,4 +1,4 @@
-package com.sparcs.teamf.dto;
+package com.sparcs.teamf.api.auth.dto;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
