+++ web/src/main/java/com/sparcs/teamf/api/question/dto/AddQuestionRequest.java
@@ -1,4 +1,4 @@
-package com.sparcs.teamf.question.dto;
+package com.sparcs.teamf.api.question.dto;
 
 public record AddQuestionRequest(String question, Long midCategoryId) {
